Move curve logic behind dedicated module
