+++ src/curve/projective.rs
@@ -6,9 +6,8 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-//! This module provides an implementation of the STARK-friendly
-//! cheetah curve over the sextic extension of the prime field Fp
-//! of characteristic p = 2^64 - 2^32 + 1.
+//! This module provides an implementation of the cheetah curve in
+//! projective coordinates.
 
 use core::{
     borrow::Borrow,
@@ -19,8 +18,9 @@ use core::{
     ops::{Add, AddAssign, Mul, MulAssign, Neg, Sub, SubAssign},
 };
 
-use crate::fp::reduce_u96;
-use crate::fp::GENERATOR;
+use super::{mul_by_3b, B};
+
+use crate::{AffinePoint, CompressedPoint, UncompressedPoint};
 use crate::{Fp, Fp6, Scalar};
 
 use crate::{MINUS_SHIFT_POINT_ARRAY, SHIFT_POINT};
@@ -34,139 +34,10 @@ use rand_core::RngCore;
 use subtle::{Choice, ConditionallySelectable, ConstantTimeEq, CtOption};
 
 #[cfg(feature = "serialize")]
-use serde::de::Visitor;
-#[cfg(feature = "serialize")]
 use serde::{self, Deserialize, Deserializer, Serialize, Serializer};
 
 impl_binops_additive!(ProjectivePoint, AffinePoint);
 impl_binops_additive_specify_output!(AffinePoint, ProjectivePoint, ProjectivePoint);
-
-//  A = 1
-/// B = u + 395
-pub const B: Fp6 = Fp6 {
-    c0: Fp(395),
-    c1: Fp::one(),
-    c2: Fp::zero(),
-    c3: Fp::zero(),
-    c4: Fp::zero(),
-    c5: Fp::zero(),
-};
-
-const B3: Fp6 = (&B).mul(&Fp6::new([3, 0, 0, 0, 0, 0]));
-
-/// An affine point
-#[derive(Copy, Clone, Debug)]
-pub struct AffinePoint {
-    pub(crate) x: Fp6,
-    pub(crate) y: Fp6,
-    pub(crate) infinity: Choice,
-}
-
-impl Default for AffinePoint {
-    fn default() -> AffinePoint {
-        AffinePoint::identity()
-    }
-}
-
-impl zeroize::DefaultIsZeroes for AffinePoint {}
-
-impl fmt::Display for AffinePoint {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{:?}", self)
-    }
-}
-
-impl Hash for AffinePoint {
-    fn hash<H: Hasher>(&self, hasher: &mut H) {
-        self.x.hash(hasher);
-        self.y.hash(hasher);
-    }
-}
-
-impl<'a> From<&'a ProjectivePoint> for AffinePoint {
-    fn from(p: &'a ProjectivePoint) -> AffinePoint {
-        let zinv = p.z.invert().unwrap_or(Fp6::zero());
-        let x = p.x * zinv;
-        let y = p.y * zinv;
-
-        let tmp = AffinePoint {
-            x,
-            y,
-            infinity: Choice::from(0u8),
-        };
-
-        AffinePoint::conditional_select(&tmp, &AffinePoint::identity(), zinv.ct_eq(&Fp6::zero()))
-    }
-}
-
-impl From<ProjectivePoint> for AffinePoint {
-    fn from(p: ProjectivePoint) -> AffinePoint {
-        AffinePoint::from(&p)
-    }
-}
-
-impl ConstantTimeEq for AffinePoint {
-    fn ct_eq(&self, other: &Self) -> Choice {
-        // The only cases in which two points are equal are
-        // 1. infinity is set on both
-        // 2. infinity is not set on both, and their coordinates are equal
-
-        (self.infinity & other.infinity)
-            | ((!self.infinity)
-                & (!other.infinity)
-                & self.x.ct_eq(&other.x)
-                & self.y.ct_eq(&other.y))
-    }
-}
-
-impl ConditionallySelectable for AffinePoint {
-    fn conditional_select(a: &Self, b: &Self, choice: Choice) -> Self {
-        AffinePoint {
-            x: Fp6::conditional_select(&a.x, &b.x, choice),
-            y: Fp6::conditional_select(&a.y, &b.y, choice),
-            infinity: Choice::conditional_select(&a.infinity, &b.infinity, choice),
-        }
-    }
-}
-
-impl Eq for AffinePoint {}
-impl PartialEq for AffinePoint {
-    #[inline]
-    fn eq(&self, other: &Self) -> bool {
-        bool::from(self.ct_eq(other))
-    }
-}
-
-impl<'a> Neg for &'a AffinePoint {
-    type Output = AffinePoint;
-
-    #[inline]
-    fn neg(self) -> AffinePoint {
-        AffinePoint {
-            x: self.x,
-            y: Fp6::conditional_select(&-self.y, &Fp6::one(), self.infinity),
-            infinity: self.infinity,
-        }
-    }
-}
-
-impl Neg for AffinePoint {
-    type Output = AffinePoint;
-
-    #[inline]
-    fn neg(self) -> AffinePoint {
-        -&self
-    }
-}
-
-impl<'a, 'b> Add<&'b ProjectivePoint> for &'a AffinePoint {
-    type Output = ProjectivePoint;
-
-    #[inline]
-    fn add(self, rhs: &'b ProjectivePoint) -> ProjectivePoint {
-        rhs.add_mixed(self)
-    }
-}
 
 impl<'a, 'b> Add<&'b AffinePoint> for &'a ProjectivePoint {
     type Output = ProjectivePoint;
@@ -174,15 +45,6 @@ impl<'a, 'b> Add<&'b AffinePoint> for &'a ProjectivePoint {
     #[inline]
     fn add(self, rhs: &'b AffinePoint) -> ProjectivePoint {
         self.add_mixed(rhs)
-    }
-}
-
-impl<'a, 'b> Sub<&'b ProjectivePoint> for &'a AffinePoint {
-    type Output = ProjectivePoint;
-
-    #[inline]
-    fn sub(self, rhs: &'b ProjectivePoint) -> ProjectivePoint {
-        self + (-rhs)
     }
 }
 
@@ -204,206 +66,6 @@ where
         I: Iterator<Item = T>,
     {
         iter.fold(Self::identity(), |acc, item| acc + item.borrow())
-    }
-}
-
-impl AffinePoint {
-    /// Returns the identity point in affine coordinates
-    pub fn identity() -> AffinePoint {
-        AffinePoint {
-            x: Fp6::zero(),
-            y: Fp6::one(),
-            infinity: Choice::from(1u8),
-        }
-    }
-
-    /// Returns the x coordinate of this AffinePoint
-    pub const fn get_x(&self) -> Fp6 {
-        self.x
-    }
-
-    /// Returns the y coordinate of this AffinePoint
-    pub const fn get_y(&self) -> Fp6 {
-        self.y
-    }
-
-    /// Computes a random `AffinePoint` element
-    pub fn random(mut rng: impl RngCore) -> Self {
-        ProjectivePoint::random(&mut rng).into()
-    }
-
-    /// Returns a fixed generator of the curve in affine coordinates
-    /// The point has been generated from the Simplified Shallue-van
-    /// de Woestijne-Ulas method for hashing to elliptic curves in
-    /// Short Weierstrass form, applied on the integer value of the
-    /// binary encoding of the string "Cheetah".
-    ///
-    /// See https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve
-    /// for more.
-    pub fn generator() -> AffinePoint {
-        AffinePoint {
-            x: Fp6 {
-                c0: Fp(0x263a588f4b0118a1),
-                c1: Fp(0x7757a0bcb26a142d),
-                c2: Fp(0x9215adfc1e925890),
-                c3: Fp(0x430aad2ce14759a4),
-                c4: Fp(0x534ece54de4b2c8),
-                c5: Fp(0xb39050f01f7b1f33),
-            },
-            y: Fp6 {
-                c0: Fp(0xd57f0d0d47482534),
-                c1: Fp(0x26821d894fa8ea0f),
-                c2: Fp(0xc77f564783ef13a1),
-                c3: Fp(0x949c360784284ec2),
-                c4: Fp(0xb7040bd639ef3cc4),
-                c5: Fp(0x8aa635f2719d255f),
-            },
-            infinity: Choice::from(0u8),
-        }
-    }
-
-    /// Outputs a compress byte representation of this `AffinePoint` element
-    pub fn to_compressed(&self) -> CompressedPoint {
-        CompressedPoint::from_affine(self)
-    }
-
-    /// Outputs an uncompressed byte representation of this `AffinePoint` element
-    /// It is twice larger than when calling `AffinePoint::to_compressed()`
-    pub fn to_uncompressed(&self) -> UncompressedPoint {
-        UncompressedPoint::from_affine(self)
-    }
-
-    /// Attempts to deserialize an uncompressed element.
-    pub fn from_uncompressed(uncompressed_point: &UncompressedPoint) -> CtOption<Self> {
-        Self::from_uncompressed_unchecked(uncompressed_point)
-            .and_then(|p| CtOption::new(p, p.is_on_curve() & p.is_torsion_free()))
-    }
-
-    /// Attempts to deserialize an uncompressed element, not checking if the
-    /// element is on the curve and not checking if it is in the correct subgroup.
-    /// **This is dangerous to call unless you trust the bytes you are reading; otherwise,
-    /// API invariants may be broken.** Please consider using `from_uncompressed()` instead.
-    pub fn from_uncompressed_unchecked(uncompressed_point: &UncompressedPoint) -> CtOption<Self> {
-        uncompressed_point.to_affine()
-    }
-
-    /// Attempts to deserialize a compressed element.
-    pub fn from_compressed(compressed_point: &CompressedPoint) -> CtOption<Self> {
-        // We already know the point is on the curve because this is established
-        // by the y-coordinate recovery procedure in from_compressed_unchecked().
-
-        Self::from_compressed_unchecked(compressed_point)
-            .and_then(|p| CtOption::new(p, p.is_torsion_free()))
-    }
-
-    /// Attempts to deserialize an uncompressed element, not checking if the
-    /// element is in the correct subgroup.
-    /// **This is dangerous to call unless you trust the bytes you are reading; otherwise,
-    /// API invariants may be broken.** Please consider using `from_compressed()` instead.
-    pub fn from_compressed_unchecked(compressed_point: &CompressedPoint) -> CtOption<Self> {
-        compressed_point.to_affine()
-    }
-
-    #[allow(unused)]
-    /// Constructs an `AffinePoint` element without checking that it is a valid point.
-    /// Assumes the coordinates do not represent the infinity point.
-    pub fn from_raw_coordinates(elems: [Fp6; 2]) -> Self {
-        AffinePoint {
-            x: elems[0],
-            y: elems[1],
-            infinity: Choice::from(0u8),
-        }
-    }
-
-    /// Returns true if this element is the identity (the point at infinity).
-    #[inline]
-    pub const fn is_identity(&self) -> Choice {
-        self.infinity
-    }
-
-    /// Returns true if this point is on the curve. This should always return
-    /// true unless an "unchecked" API was used.
-    pub fn is_on_curve(&self) -> Choice {
-        // y^2 - x^3 - x ?= B
-
-        (self.y.square() - self.x * (self.x.square() + Fp6::one())).ct_eq(&B) | self.infinity
-    }
-
-    /// Performs an affine scalar multiplication from `by`
-    /// given as byte representation of a `Scalar` element.
-    #[inline]
-    pub fn multiply(&self, by: &[u8; 32]) -> AffinePoint {
-        ProjectivePoint::multiply(&self.into(), by).into()
-    }
-
-    /// Performs an affine scalar multiplication from `by`
-    /// given as byte representation of a `Scalar` element.
-    ///
-    /// **This operation is variable time with respect
-    /// to the scalar.** If the scalar is fixed,
-    /// this operation is effectively constant time.
-    #[inline]
-    pub fn multiply_vartime(&self, by: &[u8; 32]) -> AffinePoint {
-        ProjectivePoint::multiply_vartime(&self.into(), by).into()
-    }
-
-    /// Performs the affine sum [`by_lhs` * `self` + `by_rhs` * `rhs`] with `by_lhs`
-    /// and `by_rhs` given as byte representations of `Scalar` elements.
-    #[inline]
-    pub fn multiply_double(
-        &self,
-        rhs: &AffinePoint,
-        by_lhs: &[u8; 32],
-        by_rhs: &[u8; 32],
-    ) -> AffinePoint {
-        ProjectivePoint::multiply_double(&self.into(), &rhs.into(), by_lhs, by_rhs).into()
-    }
-
-    /// Performs the affine sum [`by_lhs` * `self` + `by_rhs` * `rhs`] with `by_lhs`
-    /// and `by_rhs` given as byte representations of `Scalar` elements.
-    ///
-    /// **This operation is variable time with respect
-    /// to the scalars.** If the scalars are fixed,
-    /// this operation is effectively constant time.
-    #[inline]
-    pub fn multiply_double_vartime(
-        &self,
-        rhs: &AffinePoint,
-        by_lhs: &[u8; 32],
-        by_rhs: &[u8; 32],
-    ) -> AffinePoint {
-        ProjectivePoint::multiply_double_vartime(&self.into(), &rhs.into(), by_lhs, by_rhs).into()
-    }
-
-    /// Performs the affine sum [`by_lhs` * `self` + `by_rhs` * `rhs`] with `by_lhs`
-    /// and `by_rhs` given as byte representations of `Scalar` elements.
-    ///
-    /// **This operation is variable time with respect
-    /// to the scalars.** If the scalars are fixed,
-    /// this operation is effectively constant time.
-    #[inline]
-    pub fn multiply_double_with_basepoint_vartime(
-        &self,
-        by_self: &[u8; 32],
-        by_basepoint: &[u8; 32],
-    ) -> AffinePoint {
-        ProjectivePoint::multiply_double_with_basepoint_vartime(&self.into(), by_self, by_basepoint)
-            .into()
-    }
-
-    /// Multiplies by the curve cofactor
-    pub fn clear_cofactor(&self) -> AffinePoint {
-        let point: ProjectivePoint = self.into();
-
-        point.clear_cofactor().into()
-    }
-
-    /// Returns true if this point is free of an $h$-torsion component.
-    /// This should always return true unless an "unchecked" API was used.
-    pub fn is_torsion_free(&self) -> Choice {
-        let point: ProjectivePoint = self.into();
-
-        point.is_torsion_free()
     }
 }
 
@@ -532,72 +194,8 @@ impl<'a, 'b> Mul<&'b Scalar> for &'a ProjectivePoint {
     }
 }
 
-impl<'a, 'b> Mul<&'b Scalar> for &'a AffinePoint {
-    type Output = ProjectivePoint;
-
-    fn mul(self, other: &'b Scalar) -> Self::Output {
-        ProjectivePoint::from(self).multiply(&other.to_bytes())
-    }
-}
-
 impl_binops_additive!(ProjectivePoint, ProjectivePoint);
 impl_binops_multiplicative!(ProjectivePoint, Scalar);
-impl_binops_multiplicative_mixed!(AffinePoint, Scalar, ProjectivePoint);
-
-#[inline(always)]
-const fn mul_by_3b(a: &Fp6) -> Fp6 {
-    let aa = (&a.c0).mul(&B3.c0).0 as u128;
-    let ab = a.c0.0 as u128;
-    let ab = ab + (ab << 1);
-
-    let ba = (&a.c1).mul(&B3.c0).0 as u128;
-    let bb = a.c1.0 as u128;
-    let bb = bb + (bb << 1);
-
-    let ca = (&a.c2).mul(&B3.c0).0 as u128;
-    let cb = a.c2.0 as u128;
-    let cb = cb + (cb << 1);
-
-    let da = (&a.c3).mul(&B3.c0).0 as u128;
-    let db = a.c3.0 as u128;
-    let db = db + (db << 1);
-
-    let ea = (&a.c4).mul(&B3.c0).0 as u128;
-    let eb = a.c4.0 as u128;
-    let eb = eb + (eb << 1);
-
-    let fa = (&a.c5).mul(&B3.c0).0 as u128;
-    let fb = a.c5.0 as u128;
-    let fb = fb + (fb << 1);
-
-    let c0 = fb * GENERATOR.0 as u128;
-    let c0 = c0 + aa;
-    let c0 = Fp(reduce_u96(c0));
-
-    let c1 = ab + ba;
-    let c1 = Fp(reduce_u96(c1));
-
-    let c2 = bb + ca;
-    let c2 = Fp(reduce_u96(c2));
-
-    let c3 = cb + da;
-    let c3 = Fp(reduce_u96(c3));
-
-    let c4 = db + ea;
-    let c4 = Fp(reduce_u96(c4));
-
-    let c5 = eb + fa;
-    let c5 = Fp(reduce_u96(c5));
-
-    Fp6 {
-        c0,
-        c1,
-        c2,
-        c3,
-        c4,
-        c5,
-    }
-}
 
 impl ProjectivePoint {
     /// Returns the identity of the group: the point at infinity.
@@ -1048,8 +646,12 @@ impl ProjectivePoint {
             acc = acc.double_unchecked();
 
             match digits[j].cmp(&0) {
-                Ordering::Greater => acc += &table.get_point(digits[j] as usize),
-                Ordering::Less => acc -= &table.get_point(-digits[j] as usize),
+                Ordering::Greater => {
+                    acc = acc.add_mixed_unchecked(&table.get_point(digits[j] as usize))
+                }
+                Ordering::Less => {
+                    acc = acc.add_mixed_unchecked(&table.get_point(-digits[j] as usize).neg())
+                }
                 Ordering::Equal => (),
             };
         }
@@ -1283,218 +885,6 @@ impl ProjectivePoint {
     }
 }
 
-/// A compressed point, storing the `x` coordinate of
-/// a point, along an extra byte storing metadata
-/// to be used for decompression.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub struct CompressedPoint(pub [u8; 49]);
-
-impl Default for CompressedPoint {
-    fn default() -> Self {
-        AffinePoint::identity().to_compressed()
-    }
-}
-
-impl ConstantTimeEq for CompressedPoint {
-    fn ct_eq(&self, other: &Self) -> Choice {
-        self.as_bytes().ct_eq(other.as_bytes())
-    }
-}
-
-impl CompressedPoint {
-    /// Converts an `AffinePoint` to a `CompressedPoint`
-    fn from_affine(point: &AffinePoint) -> Self {
-        let mut bytes = [0u8; 49];
-
-        // Strictly speaking, point.x is zero already when point.infinity is true, but
-        // to guard against implementation mistakes we do not assume this.
-        bytes[0..48].copy_from_slice(
-            &Fp6::conditional_select(&point.x, &Fp6::zero(), point.infinity).to_bytes(),
-        );
-
-        // Is this point at infinity? If so, set the most significant bit of the last byte.
-        bytes[48] |= u8::conditional_select(&0u8, &(1u8 << 7), point.infinity);
-
-        // Is the y-coordinate the lexicographically largest of the two associated with the
-        // x-coordinate? If so, set the second most significant bit of the last bytes so long
-        // as this is not the point at infinity.
-        bytes[48] |= u8::conditional_select(
-            &0u8,
-            &(1u8 << 6),
-            (!point.infinity) & point.y.lexicographically_largest(),
-        );
-
-        Self(bytes)
-    }
-
-    /// Attempts to convert a `CompressedPoint` to an `AffinePoint`
-    /// The resulting point is ensured to be on the curve, but is
-    /// not necessarily on the prime order subgroup.
-    fn to_affine(self) -> CtOption<AffinePoint> {
-        // Obtain the two flags stored in the last byte of `self`
-        let infinity_flag_set = Choice::from((self.0[48] >> 7) & 1);
-        let sort_flag_set = Choice::from((self.0[48] >> 6) & 1);
-
-        // Attempt to obtain the x-coordinate
-        let x = {
-            let mut tmp = [0; 48];
-            tmp.copy_from_slice(&self.0[0..48]);
-
-            Fp6::from_bytes(&tmp)
-        };
-
-        x.and_then(|x| {
-            // If the infinity flag is set, return the value assuming
-            // the x-coordinate is zero and the sort flag is not set.
-            //
-            // Otherwise, return a recovered point (assuming the correct
-            // y-coordinate can be found) so long as the infinity flag
-            // was not set.
-            CtOption::new(
-                AffinePoint::identity(),
-                infinity_flag_set & // Infinity flag should be set
-                    (!sort_flag_set) & // Sort flag should not be set
-                    x.is_zero(), // The x-coordinate should be zero
-            )
-            .or_else(|| {
-                // Recover a y-coordinate given x by y = sqrt(x^3 + x + B)
-                ((x.square() * x) + x + B).sqrt().and_then(|y| {
-                    // Switch to the correct y-coordinate if necessary.
-                    let y = Fp6::conditional_select(
-                        &y,
-                        &-y,
-                        y.lexicographically_largest() ^ sort_flag_set,
-                    );
-
-                    CtOption::new(
-                        AffinePoint {
-                            x,
-                            y,
-                            infinity: infinity_flag_set,
-                        },
-                        !infinity_flag_set, // Infinity flag should not be set
-                    )
-                })
-            })
-        })
-    }
-
-    /// Copies the bytes of this `CompressedPoint`.
-    pub fn to_bytes(&self) -> [u8; 49] {
-        self.0
-    }
-
-    /// Views this `CompressedPoint` as an array of bytes.
-    pub fn as_bytes(&self) -> &[u8; 49] {
-        &self.0
-    }
-
-    /// Interprets the provided bytes as a `CompressedPoint`.
-    /// This does not check the validity of the input, and may result
-    /// in failure when calling `to_affine()`.
-    pub fn from_bytes(bytes: &[u8; 49]) -> Self {
-        Self(*bytes)
-    }
-}
-
-/// A uncompressed point, storing the `x` and `y` coordinates
-/// of a point, along an extra byte storing metadata to be
-/// used for decompression.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub struct UncompressedPoint(pub [u8; 97]);
-
-impl Default for UncompressedPoint {
-    fn default() -> Self {
-        AffinePoint::identity().to_uncompressed()
-    }
-}
-
-impl ConstantTimeEq for UncompressedPoint {
-    fn ct_eq(&self, other: &Self) -> Choice {
-        self.as_bytes().ct_eq(other.as_bytes())
-    }
-}
-
-impl UncompressedPoint {
-    /// Converts an `AffinePoint` to an `UncompressedPoint`
-    fn from_affine(point: &AffinePoint) -> Self {
-        let mut bytes = [0; 97];
-
-        bytes[0..48].copy_from_slice(
-            &Fp6::conditional_select(&point.x, &Fp6::zero(), point.infinity).to_bytes()[..],
-        );
-        bytes[48..96].copy_from_slice(
-            &Fp6::conditional_select(&point.y, &Fp6::zero(), point.infinity).to_bytes()[..],
-        );
-
-        // Is this point at infinity? If so, set the most significant bit of the last byte.
-        bytes[96] |= u8::conditional_select(&0u8, &(1u8 << 7), point.infinity);
-
-        Self(bytes)
-    }
-
-    /// Attempts to convert an `UncompressedPoint` to an `AffinePoint`
-    /// The resulting point is ensured to be on the curve, but is
-    /// not necessarily on the prime order subgroup.
-    fn to_affine(self) -> CtOption<AffinePoint> {
-        // Obtain the two flags
-        let infinity_flag_set = Choice::from((self.0[96] >> 7) & 1);
-
-        // Attempt to obtain the x-coordinate
-        let x = {
-            let mut tmp = [0; 48];
-            tmp.copy_from_slice(&self.0[0..48]);
-
-            Fp6::from_bytes(&tmp)
-        };
-
-        // Attempt to obtain the y-coordinate
-        let y = {
-            let mut tmp = [0; 48];
-            tmp.copy_from_slice(&self.0[48..96]);
-
-            Fp6::from_bytes(&tmp)
-        };
-
-        x.and_then(|x| {
-            y.and_then(|y| {
-                let p = AffinePoint::conditional_select(
-                    &AffinePoint {
-                        x,
-                        y,
-                        infinity: infinity_flag_set,
-                    },
-                    &AffinePoint::identity(),
-                    infinity_flag_set,
-                );
-
-                CtOption::new(
-                    p,
-                    // If the infinity flag is set, the x and y coordinates should have been zero.
-                    (!infinity_flag_set) | (infinity_flag_set & x.is_zero() & y.is_zero()),
-                )
-            })
-        })
-    }
-
-    /// Copies the bytes of this `UncompressedPoint`.
-    pub fn to_bytes(&self) -> [u8; 97] {
-        self.0
-    }
-
-    /// Views this `UncompressedPoint` as an array of bytes.
-    pub fn as_bytes(&self) -> &[u8; 97] {
-        &self.0
-    }
-
-    /// Interprets the provided bytes as a `UncompressedPoint`.
-    /// This does not check the validity of the input, and may result
-    /// in failure when calling `to_affine()`.
-    pub fn from_bytes(bytes: &[u8; 97]) -> Self {
-        Self(*bytes)
-    }
-}
-
 // GROUP TRAITS IMPLEMENTATION
 // ================================================================================================
 
@@ -1539,136 +929,12 @@ impl Curve for ProjectivePoint {
 // ================================================================================================
 
 #[cfg(feature = "serialize")]
-impl Serialize for CompressedPoint {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        use serde::ser::SerializeTuple;
-        let mut tup = serializer.serialize_tuple(49)?;
-        for byte in self.0.iter() {
-            tup.serialize_element(byte)?;
-        }
-        tup.end()
-    }
-}
-
-#[cfg(feature = "serialize")]
-impl Serialize for UncompressedPoint {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        use serde::ser::SerializeTuple;
-        let mut tup = serializer.serialize_tuple(97)?;
-        for byte in self.0.iter() {
-            tup.serialize_element(byte)?;
-        }
-        tup.end()
-    }
-}
-
-#[cfg(feature = "serialize")]
-impl Serialize for AffinePoint {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        self.to_compressed().serialize(serializer)
-    }
-}
-
-#[cfg(feature = "serialize")]
 impl Serialize for ProjectivePoint {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
     where
         S: Serializer,
     {
         self.to_compressed().serialize(serializer)
-    }
-}
-
-#[cfg(feature = "serialize")]
-impl<'de> Deserialize<'de> for CompressedPoint {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        struct CompressedPointVisitor;
-
-        impl<'de> Visitor<'de> for CompressedPointVisitor {
-            type Value = CompressedPoint;
-
-            fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-                formatter.write_str("49 bytes of data")
-            }
-
-            fn visit_seq<A>(self, mut seq: A) -> Result<CompressedPoint, A::Error>
-            where
-                A: serde::de::SeqAccess<'de>,
-            {
-                let mut bytes = [0u8; 49];
-                for (i, byte) in bytes.iter_mut().enumerate() {
-                    *byte = seq
-                        .next_element()?
-                        .ok_or_else(|| serde::de::Error::invalid_length(i, &"expected 49 bytes"))?;
-                }
-
-                Ok(CompressedPoint(bytes))
-            }
-        }
-
-        deserializer.deserialize_tuple(49, CompressedPointVisitor)
-    }
-}
-
-#[cfg(feature = "serialize")]
-impl<'de> Deserialize<'de> for UncompressedPoint {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        struct UncompressedPointVisitor;
-
-        impl<'de> Visitor<'de> for UncompressedPointVisitor {
-            type Value = UncompressedPoint;
-
-            fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-                formatter.write_str("97 bytes of data")
-            }
-
-            fn visit_seq<A>(self, mut seq: A) -> Result<UncompressedPoint, A::Error>
-            where
-                A: serde::de::SeqAccess<'de>,
-            {
-                let mut bytes = [0u8; 97];
-                for (i, byte) in bytes.iter_mut().enumerate() {
-                    *byte = seq
-                        .next_element()?
-                        .ok_or_else(|| serde::de::Error::invalid_length(i, &"expected 97 bytes"))?;
-                }
-
-                Ok(UncompressedPoint(bytes))
-            }
-        }
-
-        deserializer.deserialize_tuple(97, UncompressedPointVisitor)
-    }
-}
-
-#[cfg(feature = "serialize")]
-impl<'de> Deserialize<'de> for AffinePoint {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let compressed_point = CompressedPoint::deserialize(deserializer)?;
-        let p = AffinePoint::from_compressed(&compressed_point);
-        if bool::from(p.is_some()) {
-            Ok(p.unwrap())
-        } else {
-            Err(serde::de::Error::custom("decompression failed"))
-        }
     }
 }
 
@@ -1697,8 +963,6 @@ mod tests {
 
     #[test]
     fn test_is_on_curve() {
-        assert!(bool::from(AffinePoint::identity().is_on_curve()));
-        assert!(bool::from(AffinePoint::generator().is_on_curve()));
         assert!(bool::from(ProjectivePoint::identity().is_on_curve()));
         assert!(bool::from(ProjectivePoint::generator().is_on_curve()));
 
@@ -1722,23 +986,6 @@ mod tests {
 
         test.x = z;
         assert!(!bool::from(test.is_on_curve()));
-    }
-
-    #[test]
-    #[allow(clippy::eq_op)]
-    fn test_affine_point_equality() {
-        let a = AffinePoint::generator();
-        let b = AffinePoint::identity();
-        let c = AffinePoint::default();
-
-        assert!(a == a);
-        assert!(b == b);
-        assert!(b == c);
-        assert!(a != b);
-        assert!(b != a);
-
-        assert!(bool::from(b.is_identity()));
-        assert!(!bool::from(a.ct_eq(&b)));
     }
 
     #[test]
@@ -2017,16 +1264,6 @@ mod tests {
     }
 
     #[test]
-    fn test_affine_negation_and_subtraction() {
-        let a = AffinePoint::generator();
-        assert_eq!(ProjectivePoint::from(a) + (-a), ProjectivePoint::identity());
-        assert_eq!(
-            ProjectivePoint::from(a) + (-a),
-            ProjectivePoint::from(a) - a
-        );
-    }
-
-    #[test]
     fn test_projective_scalar_multiplication() {
         let mut rng = OsRng;
         let g = ProjectivePoint::generator();
@@ -2091,43 +1328,6 @@ mod tests {
     }
 
     #[test]
-    fn test_affine_scalar_multiplication() {
-        let mut rng = OsRng;
-        let g = AffinePoint::generator();
-
-        for _ in 0..100 {
-            let a = Scalar::random(&mut rng);
-            let b = Scalar::random(&mut rng);
-
-            let c = a * b;
-
-            assert_eq!((g * a) * b, g * c);
-            assert_eq!(g.multiply_vartime(&c.to_bytes()), g.multiply(&c.to_bytes()));
-        }
-    }
-
-    #[test]
-    fn test_affine_double_scalar_multiplication() {
-        let mut rng = OsRng;
-        let g = AffinePoint::from(ProjectivePoint::generator() * Scalar::random(&mut rng));
-        let h = AffinePoint::from(ProjectivePoint::generator() * Scalar::random(&mut rng));
-
-        for _ in 0..100 {
-            let a = Scalar::random(&mut rng);
-            let b = Scalar::random(&mut rng);
-
-            assert_eq!(
-                g.multiply_double(&h, &a.to_bytes(), &b.to_bytes()),
-                AffinePoint::from(g * a + h * b)
-            );
-            assert_eq!(
-                g.multiply_double_vartime(&h, &a.to_bytes(), &b.to_bytes()),
-                AffinePoint::from(g * a + h * b)
-            );
-        }
-    }
-
-    #[test]
     fn test_clear_cofactor() {
         // the generator (and the identity) are always on the curve
         let generator = ProjectivePoint::generator();
@@ -2185,10 +1385,6 @@ mod tests {
             },
             infinity: Choice::from(0u8),
         };
-        assert!(bool::from(a.is_on_curve()));
-        assert!(!bool::from(a.is_torsion_free()));
-        assert!(bool::from(AffinePoint::identity().is_torsion_free()));
-        assert!(bool::from(AffinePoint::generator().is_torsion_free()));
 
         let a = ProjectivePoint::from(&a);
         assert!(bool::from(a.is_on_curve()));
@@ -2240,10 +1436,6 @@ mod tests {
     fn test_zeroize() {
         use zeroize::Zeroize;
 
-        let mut a = AffinePoint::generator();
-        a.zeroize();
-        assert_eq!(a, AffinePoint::identity());
-
         let mut a = ProjectivePoint::generator();
         a.zeroize();
         assert_eq!(a, ProjectivePoint::identity());
@@ -2257,11 +1449,6 @@ mod tests {
         let mut rng = OsRng;
         // Random points
         for _ in 0..100 {
-            let point = AffinePoint::random(&mut rng);
-            let bytes = point.to_compressed();
-            let point_decompressed = AffinePoint::from_compressed(&bytes).unwrap();
-            assert_eq!(point, point_decompressed);
-
             let point = ProjectivePoint::random(&mut rng);
             let bytes = point.to_compressed();
             let point_decompressed = ProjectivePoint::from_compressed(&bytes).unwrap();
@@ -2270,10 +1457,6 @@ mod tests {
 
         // Identity point
         {
-            let bytes = AffinePoint::identity().to_compressed();
-            let point_decompressed = AffinePoint::from_compressed(&bytes).unwrap();
-            assert!(bool::from(point_decompressed.is_identity()));
-
             let bytes = ProjectivePoint::identity().to_compressed();
             let point_decompressed = ProjectivePoint::from_compressed(&bytes).unwrap();
             assert!(bool::from(point_decompressed.is_identity()));
@@ -2294,18 +1477,10 @@ mod tests {
                 y: Fp6::zero(),
                 infinity: 0.into(),
             };
-            let bytes = point.to_compressed();
-            let point_decompressed = AffinePoint::from_compressed(&bytes);
-            assert!(bool::from(point_decompressed.is_none()));
 
             let point = ProjectivePoint::from(&point);
             let bytes = point.to_compressed();
             let point_decompressed = ProjectivePoint::from_compressed(&bytes);
-            assert!(bool::from(point_decompressed.is_none()));
-        }
-        {
-            let bytes = CompressedPoint([255; 49]);
-            let point_decompressed = AffinePoint::from_compressed_unchecked(&bytes);
             assert!(bool::from(point_decompressed.is_none()));
         }
     }
@@ -2316,11 +1491,6 @@ mod tests {
 
         // Random points
         for _ in 0..100 {
-            let point = AffinePoint::random(&mut rng);
-            let bytes = point.to_uncompressed();
-            let point_decompressed = AffinePoint::from_uncompressed(&bytes).unwrap();
-            assert_eq!(point, point_decompressed);
-
             let point = ProjectivePoint::random(&mut rng);
             let bytes = point.to_uncompressed();
             let point_decompressed = ProjectivePoint::from_uncompressed(&bytes).unwrap();
@@ -2329,10 +1499,6 @@ mod tests {
 
         // Identity point
         {
-            let bytes = AffinePoint::identity().to_uncompressed();
-            let point_decompressed = AffinePoint::from_uncompressed(&bytes).unwrap();
-            assert!(bool::from(point_decompressed.is_identity()));
-
             let bytes = ProjectivePoint::identity().to_uncompressed();
             let point_decompressed = ProjectivePoint::from_uncompressed(&bytes).unwrap();
             assert!(bool::from(point_decompressed.is_identity()));
@@ -2355,9 +1521,6 @@ mod tests {
                 y: Fp6::zero(),
                 infinity: 0.into(),
             };
-            let bytes = point.to_uncompressed();
-            let point_decompressed = AffinePoint::from_uncompressed(&bytes);
-            assert!(bool::from(point_decompressed.is_none()));
 
             let point = ProjectivePoint::from(&point);
             let bytes = point.to_uncompressed();
@@ -2372,8 +1535,6 @@ mod tests {
                 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
                 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
             ]);
-            let point_decompressed = AffinePoint::from_uncompressed_unchecked(&bytes);
-            assert!(bool::from(point_decompressed.is_none()));
 
             let point_decompressed = ProjectivePoint::from_uncompressed_unchecked(&bytes);
             assert!(bool::from(point_decompressed.is_none()));
@@ -2386,8 +1547,6 @@ mod tests {
                 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255,
                 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255,
             ]);
-            let point_decompressed = AffinePoint::from_uncompressed_unchecked(&bytes);
-            assert!(bool::from(point_decompressed.is_none()));
 
             let point_decompressed = ProjectivePoint::from_uncompressed_unchecked(&bytes);
             assert!(bool::from(point_decompressed.is_none()));
@@ -2396,32 +1555,6 @@ mod tests {
 
     // SERDE SERIALIZATIOIN
     // ================================================================================================
-
-    #[test]
-    #[cfg(feature = "serialize")]
-    fn test_serde_affine() {
-        let mut rng = OsRng;
-        let point = AffinePoint::random(&mut rng);
-        let encoded = bincode::serialize(&point).unwrap();
-        let parsed: AffinePoint = bincode::deserialize(&encoded).unwrap();
-        assert_eq!(parsed, point);
-
-        // Check that the encoding is 49 bytes exactly
-        assert_eq!(encoded.len(), 49);
-
-        // Check that the encoding itself matches the usual one
-        assert_eq!(
-            point,
-            bincode::deserialize(&point.to_compressed().0).unwrap()
-        );
-
-        // Check that invalid encodings fail
-        let encoded = bincode::serialize(&point).unwrap();
-        assert!(bincode::deserialize::<AffinePoint>(&encoded[0..47]).is_err());
-
-        let encoded = [255; 49];
-        assert!(bincode::deserialize::<AffinePoint>(&encoded).is_err());
-    }
 
     #[test]
     #[cfg(feature = "serialize")]
